rewrite formatters to ext classes
