+++ sources/awesomewidgets/awabstractformatter.cpp
@@ -17,23 +17,16 @@
 
 #include "awabstractformatter.h"
 
-#include <QDir>
 #include <QSettings>
-#include <QStandardPaths>
 
 #include "awdebug.h"
 
 
 AWAbstractFormatter::AWAbstractFormatter(QWidget *parent,
                                          const QString filePath)
-    : QDialog(parent)
-    , m_fileName(filePath)
+    : AbstractExtItem(parent, filePath)
 {
     qCDebug(LOG_AW) << __PRETTY_FUNCTION__;
-
-    qCDebug(LOG_AW) << "Desktop name" << filePath;
-
-    m_name = m_fileName;
 }
 
 
@@ -43,66 +36,23 @@ AWAbstractFormatter::~AWAbstractFormatter()
 }
 
 
-void AWAbstractFormatter::copyDefaults(AWAbstractFormatter *_other) const
+void AWAbstractFormatter::copyDefaults(AbstractExtItem *_other) const
 {
-    _other->setComment(m_comment);
-    _other->setName(m_name);
-    _other->setType(m_type);
+    AbstractExtItem::copyDefaults(_other);
+
+    static_cast<AWAbstractFormatter *>(_other)->setType(m_type);
 }
 
 
-QString AWAbstractFormatter::writtableConfig() const
+QString AWAbstractFormatter::uniq() const
 {
-    QStringList paths = m_fileName.split(QChar('/'));
-
-    QString name = paths.takeLast();
-    QString dir = paths.takeLast();
-
-    return QString("%1/awesomewidgets/%2/%3")
-        .arg(QStandardPaths::writableLocation(
-            QStandardPaths::GenericDataLocation))
-        .arg(dir)
-        .arg(name);
-}
-
-
-QString AWAbstractFormatter::comment() const
-{
-    return m_comment;
-}
-
-
-QString AWAbstractFormatter::fileName() const
-{
-    return m_fileName;
-}
-
-
-QString AWAbstractFormatter::name() const
-{
-    return m_name;
+    return QString("%1(%2)").arg(name()).arg(m_type);
 }
 
 
 QString AWAbstractFormatter::type() const
 {
     return m_type;
-}
-
-
-void AWAbstractFormatter::setComment(const QString _comment)
-{
-    qCDebug(LOG_AW) << "Comment" << _comment;
-
-    m_comment = _comment;
-}
-
-
-void AWAbstractFormatter::setName(const QString _name)
-{
-    qCDebug(LOG_AW) << "Name" << _name;
-
-    m_name = _name;
 }
 
 
@@ -116,34 +66,24 @@ void AWAbstractFormatter::setType(const QString _type)
 
 void AWAbstractFormatter::readConfiguration()
 {
-    QSettings settings(m_fileName, QSettings::IniFormat);
+    AbstractExtItem::readConfiguration();
+
+    QSettings settings(fileName(), QSettings::IniFormat);
 
     settings.beginGroup(QString("Desktop Entry"));
-    setName(settings.value(QString("Name"), m_name).toString());
-    setComment(settings.value(QString("Comment"), m_comment).toString());
     setType(settings.value(QString("Type"), m_type).toString());
     settings.endGroup();
 }
 
 
-bool AWAbstractFormatter::tryDelete() const
-{
-    bool status = QFile::remove(m_fileName);
-    qCInfo(LOG_AW) << "Remove file" << m_fileName << status;
-
-    return status;
-}
-
-
 void AWAbstractFormatter::writeConfiguration() const
 {
+    AbstractExtItem::writeConfiguration();
+
     QSettings settings(writtableConfig(), QSettings::IniFormat);
     qCInfo(LOG_AW) << "Configuration file" << settings.fileName();
 
     settings.beginGroup(QString("Desktop Entry"));
-    settings.setValue(QString("Encoding"), QString("UTF-8"));
-    settings.setValue(QString("Name"), m_name);
-    settings.setValue(QString("Comment"), m_comment);
     settings.setValue(QString("Type"), m_type);
     settings.endGroup();
