+++ sources/awesomewidgets/awfloatformatter.ui
