+++ sources/awesomewidgets/awnoformatter.ui
